deleteOwner ACTUALLY passing the npm test

+++ Controllers/PetOwnersController.cs
@@ -34,13 +34,20 @@ namespace pet_hotel.Controllers
         }
 
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public IActionResult Delete(int id)
         {
             PetOwner petOwner = _context.PetOwners.Find(id);
+
+            if(petOwner is null)
+            {
+                return NotFound();
+            }
 
             _context.PetOwners.Remove(petOwner);
 
             _context.SaveChanges();
+
+            return NoContent();
         }
     }
 }
